HW_08. Added checking for file existence in AbstractFileStorage.clear().

+++ src/storage/AbstractFileStorage.java
@@ -24,7 +24,9 @@ public abstract class AbstractFileStorage extends AbstractStorage<File> {
     @Override
     public void clear() {
         for (File file : Objects.requireNonNull(directory.listFiles())) {
+            if(file.exists()){
             file.delete();
+            }
         }
     }
